requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
asyncpg>=0.27.0
google-generativeai>=0.3.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
python-dotenv>=1.0.0
httpx>=0.26.0
boto3>=1.28.0
greenlet>=3.0.0
aiosqlite>=0.19.0
